SB-69: Refactor Courts Lists View
