+++ app/(admin)/courts/[id]/CourtSettingsView.tsx
@@ -2,12 +2,12 @@
 
 import { useEffect, useState } from 'react';
 import { toast } from 'react-toastify';
-import hasErrorMessage from '../../utils/Error/ErrorHelper';
+import hasErrorMessage from '../../../utils/Error/ErrorHelper';
 import CourtSettings, {
 	Court,
-} from '../../components/Courts/CourtSettings/CourtSettings';
-import { getCourt, editCourt } from '../../firebase/courts/courts';
-import Loading from '../../components/UI/Loading/Loading';
+} from '../../../components/Courts/CourtSettings/CourtSettings';
+import { getCourt, editCourt } from '../../../firebase/courts/courts';
+import Loading from '../../../components/UI/Loading/Loading';
 
 type Props = {
 	courtId: string;
