+++ app/(admin)/courts/page.tsx
@@ -3,26 +3,23 @@
 import { useEffect, useState } from 'react';
 import { useRouter } from 'next/navigation';
 import { toast } from 'react-toastify';
-import Link from 'next/link';
-import { useAuthContext } from '../../context/AuthContext';
+import CourtCard from '@/app/components/Courts/CourtCard';
 import Loading from '../../components/UI/Loading/Loading';
 import hasErrorMessage from '../../utils/Error/ErrorHelper';
 import { Court } from '../../components/Courts/CourtSettings/CourtSettings';
 import { getAllCourts } from '../../firebase/courts/courts';
 
 export default function CourtsView() {
-	const user = useAuthContext();
 	const router = useRouter();
 	const [courts, setCourts] = useState<Array<Court>>([]);
+	const [loading, setLoading] = useState<boolean>(true);
 
 	useEffect((): void => {
-		if (!user) {
-			router.push('/signin');
-		}
 		const fetchData = async () => {
 			try {
 				const courtsData = await getAllCourts();
 				setCourts(courtsData);
+				setLoading(false);
 			} catch (error: unknown) {
 				if (hasErrorMessage(error)) {
 					toast.error(error.message, { theme: 'colored' });
@@ -30,22 +27,15 @@ export default function CourtsView() {
 			}
 		};
 		fetchData();
-	}, [user, router]);
+	}, [router]);
 
-	return !user ? (
+	return loading ? (
 		<Loading />
 	) : (
-		<>
-			<h2>List of courts</h2>
-			<ul>
-				{courts.map(court => (
-					<li key={court.id}>
-						<Link href={`/courts/${court.id}`}>
-							{court.id} - {court.name}
-						</Link>
-					</li>
-				))}
-			</ul>
-		</>
+		<div className='d-flex gap-3 flex-wrap flex-sm-row flex-column align-items-center justify-content-around'>
+			{courts.map(court => (
+				<CourtCard key={court.id} name={court.name} id={court.id} />
+			))}
+		</div>
 	);
 }
